refactor: basic question box 로직 분리 #15
